Move TableCommands into its own file
